requirements.txt
# Core libraries
numpy>=1.19
opencv-python>=4.5.0
PyYAML>=5.3.1
requests>=2.25.0

# Deep learning
torch>=1.9.0
torchvision>=0.10.0
onnxruntime>=1.8.0

# Scientific computing
scipy>=1.5.0
